Refactor main
Split the main `run` function into one that gets the inputs from action and one that does the actual work. This way we don't have to mock around with inputs in the tests.

Also rename everything in main from `plugin` to `addon`, since this is part of an addon-loader, not plugin-loader :)