Add workflow to run `npm run bundle` after dependabot updates
The new `dependabot-bundle.yaml` workflow runs on dependabot PRs and commits changes to the repo after running `npm run bundle`.
It only runs when the PR has a single commit, to prevent infinite recursion or updating anything added manually.

It requires a `DEPENDABOT_BUNDLE_TOKEN` secret with a GitHub token with `contents: write` permissions to run.